requirement.txt
django==5.0.4
djangorestframework==3.15.1
pymysql==1.1.0
django-grappelli==4.0.1